+++ software/dac.h
@@ -2,14 +2,13 @@
  * dac.h
  *
  *  Created on: 24.06.2014
- *      Author: Stefan Urban
+ *      Author: Stefan Urban <stefan.urban@live.de>
  */
 
 #ifndef DAC_H_
 #define DAC_H_
 
 #include <stdint.h>
-#include <stdbool.h>
 
 #include "microchip_mcp4901.h"
 
@@ -23,18 +22,18 @@ typedef struct {
 	/**
 	 * Function pointers
 	 */
-	void (*set_value_function)(uint16_t new_dac_value);
-	uint16_t (*get_value_function)(void);
+	void (*set_value_function)(int new_dac_value);
+	int (*get_value_function)(void);
 	
 	/**
 	 * Reference voltage
 	 */
-	uint16_t reference_voltage;
+	int reference_voltage;
 	
 	/**
 	 * Resolution
 	 */
-	uint8_t resolution;
+	int resolution;
 
 } dac_t;
 
@@ -64,7 +63,7 @@ void dac_disable(dac_t dac);
  *
  * @param new_voltage
  */
-void dac_set_voltage(dac_t dac, uint16_t new_voltage);
+void dac_set_voltage(dac_t dac, int new_voltage);
 
 
 /**
@@ -72,7 +71,7 @@ void dac_set_voltage(dac_t dac, uint16_t new_voltage);
  *
  * @return voltage, in mV
  */
-uint16_t dac_get_voltage(dac_t dac);
+int dac_get_voltage(dac_t dac);
